Fixed Root Deserialize
When the root tag is TAG_End there is no NBT data. In this case it is necessary to return an empty Compound because it is impossible to tell if the root tag is TAG_End while preserving the reader before passing it to the deserializer. There is an example of this   in the NBT data for the Slot data structure, where no NBT data is represented as TAG_End. https://wiki.vg/Slot_Data